requirements.txt
Flask
flask-cors
python-dotenv
requests
together
werkzeug
numpy
joblib
xgboost
pdfplumber
python-docx
gunicorn